Make dataset lock recursive
Previously we had methods called within the scope of locks, but this
already broke as we forgot locks in other callers.

With recursive lock we can add locks directly where we need them.
Note that dataset does not use read locks and also shall never use them
as it may not obtain read+write at the same time acc. to Qt docs:
> Note that the lock type cannot be changed when trying to lock recursively, i.e. it is not possible to lock for reading in a thread that already has locked for writing (and vice versa).