Refactor viewer interface/base class
Let base handle datastates. Avoid code duplication and make interface
in derived classes more intuitive.

Also carry over proper reset of QChartViews on scene loss; as QChartView
assumes ownership of a chart, it doesn't realize when a chart gets
deleted. This is in contrast to QGraphicsView, which will realize when
its scene is deleted and reset itself properly.