Untangle Viewer from QMainWindow
Allow viewers to be anything, even not a widget at all.
For convenience, the Viewer class holds a QWidget*, but it need not be
used. The regular use-case is to have a widget, which means that the
viewer has the widget as its parent. This means the lifetime to the view
is tied to the lifetime of the widget.

Drawback is, Viewer-derived classes cannot be used as promoted widgets
in Designer anymore.