remove the crappy old 'avoid double compute' logic
It was buggy, maybe ok at the end but seemed still not to cover all
cases

now we expect frontend to properly handle this (not cram lots of ms jobs
into the queue). a window has a single control now that keeps things
clean. Although we could allow enqueueing of several configurations
after another, it is too confusing to the user anyways.
If the user likes to mess with it by using two windows on the
same dataset at once, any newer computation trigger will simply cancel
(and wait) the old one.

We could also allow several instances of meanshift at once per dataset
but there is no good use-case and it is also not a good idea as ms is
highly parallelized.